requirements.txt
pyqrcode
PyPNG
